add hound file.
test hound config

point to the javascriptfile

try to fix hound error

try the basic config

try ignore file.

fix the json layout

add react and require to the style.

try to move the parser to a sub directory.

rename and move the parsers

update the line length parser./

fix the errors i introduced.  to the components