DEC-730 - Node Deployment work.
 * DEC-741 Migrate Site index page
 * DEC-742 Migrate Collection index page
 * DEC-743 Migrate Collection introduction page
 * DEC-744 Migrate Showcase page
 * DEC-745 Migrate Search page
 * DEC-746 Setup Flux store

Partially complete:

 * DEC-752 Migrate Pages page
 * DEC-751 Fix depreciation warnings

This commit removes any dependency on Ruby. To start a new server:

`
node ./server.js
`

To create at development build:

`
npm run build-dev
`

FYI: it is probably easier to delete your current Beehive installation and clone new from this branch. Don't forget to run:

`
npm install
`

This build of Beehive is based on material-ui-prepack:

 * https://github.com/ipselon/material-ui-prepack

There are still some things that need to be removed and some additional work to be completed for a smooth node pipeline.