DEC-785 Backport multi-facet stuff to Beehive
 * Backported from Zucchetto
 * Also should resolve DEC-786 First url on Search page does not pass parameters
   back to Honeycomb search query.

Basically, facets are now being stored in an array instead of a single object.
The majority of changes are related to this change by changing parameter types
and making sure to iterate through the array where appropriate.

update .gitignore

hound

hound

fix stuff that hounds said was broken but wasnt